requirements.txt
Django==1.11
Fabric3==1.13.1.post1
Jinja2==2.9.6
MarkupSafe==1.0
PyYAML==3.12
aiodns==1.1.1
aiohttp==2.2.5
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
asn1crypto==0.22.0
async-timeout==1.3.0
billiard==3.3.0.23
cchardet==2.1.1
celery==3.1.23
cffi==1.10.0
chardet==3.0.4
cryptography==2.0.3
diff-match-patch==20121119
dj-database-url==0.3.0
django-extensions==1.7.8
django-redis==4.6.0
djangorestframework==3.5.3
funcy==1.7.3
gunicorn==19.3.0
idna==2.5
kombu==3.0.35
multidict==3.1.3
packaging==16.8
paramiko==2.1.2
psycopg2==2.6.1
pyasn1==0.2.3
pycares==2.3.0
pycparser==2.17
pycrypto==2.6.1
pylibscrypt==1.6.0
pyparsing==2.2.0
python-frontmatter==0.4.2
pytz==2017.2
redis==2.10.5
requests==2.13.0
scrypt==0.8.0
six==1.10.0
yarl==0.12.0

# NSFW image recognition
tensorflow==1.3.0
scikit-image==0.13.0
